updated test 2, 3
