+++ tests/searchvalidinput.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import SearchPage from "../pages/homePage/searchPage";
+import SearchPage from "../pages/searchPage/searchPage";
 
 test("Verify Searching" , async ({ page }) => {
     const searchPage = new SearchPage(page);
@@ -7,4 +7,6 @@ test("Verify Searching" , async ({ page }) => {
     await searchPage.clickSearchEngine();
     await searchPage.fillSerachBar("mobitel");
     await searchPage.submitSearch();
+    await searchPage.verifySearchValidInput();
+    
 });
